[fix](cloud) Prevent MS throttle timer overflow
### What problem does this PR solve?

Issue Number: N/A

Related PR: #66969, #66940

Problem Summary: The table-level MS backpressure coordinator accumulated two signed 32-bit millisecond counters after the first MS_BUSY event. At 1000 ticks per second, they could overflow after about 24.85 days and make upgrade cooldown or downgrade timing incorrect. Use 64-bit counters with overflow-free saturation at each decision threshold. Stop the MS_BUSY timer when no upgrade history is pending, while preserving the upgrade cooldown by saturating rather than resetting its counter.

### Release note

Fix long-running BE table-level MS backpressure timing so coordinator counters cannot overflow.

### Check List (For Author)

- Test: Unit Test
    - `./run-be-ut.sh --run --filter='MSBackpressureHandlerTest.*:RpcThrottleStateMachineTest.*:RpcThrottleCoordinatorTest.*:RpcThrottleIntegrationTest.*' -j100` (44 tests passed)
    - `./build.sh --be -j100`
    - `build-support/check-format.sh`
    - `build-support/check-build-hygiene.sh`
- Behavior changed: Yes. Coordinator counters saturate at timing thresholds and the MS_BUSY timer is inactive without pending downgrade history.
- Does this need documentation: No

+++ be/src/cloud/cloud_ms_backpressure_handler.cpp
@@ -412,13 +412,13 @@ MSBackpressureHandler::~MSBackpressureHandler() {
 
 void MSBackpressureHandler::_tick_thread_callback() {
     // Fixed tick interval: 1 second. Since 1 tick = 1 ms, advance by 1000 ticks each iteration.
-    constexpr int kTickIntervalMs = 1000;
+    constexpr int64_t kTickIntervalMs = 1000;
     while (!_stop_latch.wait_for(std::chrono::milliseconds(kTickIntervalMs))) {
         _advance_time(kTickIntervalMs);
     }
 }
 
-void MSBackpressureHandler::_advance_time(int ticks) {
+void MSBackpressureHandler::_advance_time(int64_t ticks) {
     if (!config::enable_ms_backpressure_handling &&
         !config::enable_ms_backpressure_handling_dry_run) {
         return;
@@ -519,11 +519,11 @@ size_t MSBackpressureHandler::upgrade_level() const {
     return _state_machine->upgrade_level();
 }
 
-int MSBackpressureHandler::ticks_since_last_ms_busy() const {
+int64_t MSBackpressureHandler::ticks_since_last_ms_busy() const {
     return _coordinator->ticks_since_last_ms_busy();
 }
 
-int MSBackpressureHandler::ticks_since_last_upgrade() const {
+int64_t MSBackpressureHandler::ticks_since_last_upgrade() const {
     return _coordinator->ticks_since_last_upgrade();
 }
